requirements.txt
python-docx>=0.8.11
PyMuPDF>=1.19.0
faiss-cpu>=1.7.4
pandas>=1.5.0
openpyxl>=3.1.0
streamlit>=1.24.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.2.0+cpu
huggingface_hub>=0.16.4
protobuf>=3.20.0
tokenizers>=0.13.3
transformers>=4.31.0
sentence-transformers>=2.2.2